Update the butterworth example
Add and use LMH6624 OP-amp in butterworth schema

Xyce gives a "timestamp too small" error with LMV681.
Use LMH6624 instead and clean up the schematic a bit.

Start using a prerendered HTML page as an example of the notebook
use instead of keeping the data in the notebook.

Move the notebook and results into the examples/butterworth directory.